Split serialize function in their own files
