+++ packages/rum/src/domain/record/index.ts
@@ -1,4 +1,4 @@
 export { record } from './record'
-export { serializeNodeWithId, serializeDocument, SerializationContextStatus } from './serializers/serialize'
+export { serializeNodeWithId, serializeDocument, SerializationContextStatus } from './serialization'
 export { createElementsScrollPositions } from './elementsScrollPositions'
 export { ShadowRootsController } from './shadowRootsController'
