+++ packages/rum/src/domain/record/serialization/getAttributesForPrivacyLevel.ts
@@ -1,0 +1,151 @@
+import { NodePrivacyLevel } from '../../../constants'
+import { shouldMaskNode } from '../privacy'
+import { getElementInputValue, switchToAbsoluteUrl, getValidTagName } from './serializationUtils'
+import type { SerializeOptions } from './serialization.types'
+import { SerializationContextStatus } from './serialization.types'
+import { serializeAttribute } from './serializeAttribute'
+
+export function getAttributesForPrivacyLevel(
+  element: Element,
+  nodePrivacyLevel: NodePrivacyLevel,
+  options: SerializeOptions
+): Record<string, string | number | boolean> {
+  if (nodePrivacyLevel === NodePrivacyLevel.HIDDEN) {
+    return {}
+  }
+  const safeAttrs: Record<string, string | number | boolean> = {}
+  const tagName = getValidTagName(element.tagName)
+  const doc = element.ownerDocument
+
+  type HtmlAttribute = { name: string; value: string }
+  for (let i = 0; i < element.attributes.length; i += 1) {
+    const attribute = element.attributes.item(i) as HtmlAttribute
+    const attributeName = attribute.name
+    const attributeValue = serializeAttribute(element, nodePrivacyLevel, attributeName, options.configuration)
+    if (attributeValue !== null) {
+      safeAttrs[attributeName] = attributeValue
+    }
+  }
+
+  if (
+    (element as HTMLInputElement).value &&
+    (tagName === 'textarea' || tagName === 'select' || tagName === 'option' || tagName === 'input')
+  ) {
+    const formValue = getElementInputValue(element, nodePrivacyLevel)
+    if (formValue !== undefined) {
+      safeAttrs.value = formValue
+    }
+  }
+
+  /**
+   * <Option> can be selected, which occurs if its `value` matches ancestor `<Select>.value`
+   */
+  if (tagName === 'option' && nodePrivacyLevel === NodePrivacyLevel.ALLOW) {
+    // For privacy=`MASK`, all the values would be the same, so skip.
+    const optionElement = element as HTMLOptionElement
+    if (optionElement.selected) {
+      safeAttrs.selected = optionElement.selected
+    }
+  }
+
+  // remote css
+  if (tagName === 'link') {
+    const stylesheet = Array.from(doc.styleSheets).find((s) => s.href === (element as HTMLLinkElement).href)
+    const cssText = getCssRulesString(stylesheet)
+    if (cssText && stylesheet) {
+      safeAttrs._cssText = cssText
+    }
+  }
+
+  // dynamic stylesheet
+  if (
+    tagName === 'style' &&
+    (element as HTMLStyleElement).sheet &&
+    // TODO: Currently we only try to get dynamic stylesheet when it is an empty style element
+    !((element as HTMLStyleElement).innerText || element.textContent || '').trim().length
+  ) {
+    const cssText = getCssRulesString((element as HTMLStyleElement).sheet)
+    if (cssText) {
+      safeAttrs._cssText = cssText
+    }
+  }
+
+  /**
+   * Forms: input[type=checkbox,radio]
+   * The `checked` property for <input> is a little bit special:
+   * 1. el.checked is a setter that returns if truthy.
+   * 2. getAttribute returns the string value
+   * getAttribute('checked') does not sync with `Element.checked`, so use JS property
+   * NOTE: `checked` property exists on `HTMLInputElement`. For serializer assumptions, we check for type=radio|check.
+   */
+  const inputElement = element as HTMLInputElement
+  if (tagName === 'input' && (inputElement.type === 'radio' || inputElement.type === 'checkbox')) {
+    if (nodePrivacyLevel === NodePrivacyLevel.ALLOW) {
+      safeAttrs.checked = !!inputElement.checked
+    } else if (shouldMaskNode(inputElement, nodePrivacyLevel)) {
+      delete safeAttrs.checked
+    }
+  }
+
+  /**
+   * Serialize the media playback state
+   */
+  if (tagName === 'audio' || tagName === 'video') {
+    const mediaElement = element as HTMLMediaElement
+    safeAttrs.rr_mediaState = mediaElement.paused ? 'paused' : 'played'
+  }
+
+  /**
+   * Serialize the scroll state for each element only for full snapshot
+   */
+  let scrollTop: number | undefined
+  let scrollLeft: number | undefined
+  const serializationContext = options.serializationContext
+  switch (serializationContext.status) {
+    case SerializationContextStatus.INITIAL_FULL_SNAPSHOT:
+      scrollTop = Math.round(element.scrollTop)
+      scrollLeft = Math.round(element.scrollLeft)
+      if (scrollTop || scrollLeft) {
+        serializationContext.elementsScrollPositions.set(element, { scrollTop, scrollLeft })
+      }
+      break
+    case SerializationContextStatus.SUBSEQUENT_FULL_SNAPSHOT:
+      if (serializationContext.elementsScrollPositions.has(element)) {
+        ;({ scrollTop, scrollLeft } = serializationContext.elementsScrollPositions.get(element)!)
+      }
+      break
+  }
+  if (scrollLeft) {
+    safeAttrs.rr_scrollLeft = scrollLeft
+  }
+  if (scrollTop) {
+    safeAttrs.rr_scrollTop = scrollTop
+  }
+
+  return safeAttrs
+}
+
+function getCssRulesString(cssStyleSheet: CSSStyleSheet | undefined | null): string | null {
+  if (!cssStyleSheet) {
+    return null
+  }
+  let rules: CSSRuleList | undefined
+  try {
+    rules = cssStyleSheet.rules || cssStyleSheet.cssRules
+  } catch {
+    // if css is protected by CORS we cannot access cssRules see: https://www.w3.org/TR/cssom-1/#the-cssstylesheet-interface
+  }
+  if (!rules) {
+    return null
+  }
+  const styleSheetCssText = Array.from(rules, getCssRuleString).join('')
+  return switchToAbsoluteUrl(styleSheetCssText, cssStyleSheet.href)
+}
+
+function getCssRuleString(rule: CSSRule): string {
+  return isCSSImportRule(rule) ? getCssRulesString(rule.styleSheet) || '' : rule.cssText
+}
+
+function isCSSImportRule(rule: CSSRule): rule is CSSImportRule {
+  return 'styleSheet' in rule
+}
