+++ packages/rum/src/domain/record/serialization/index.ts
@@ -1,0 +1,10 @@
+export {
+  getElementInputValue,
+  getSerializedNodeId,
+  hasSerializedNode,
+  nodeAndAncestorsHaveSerializedNode,
+} from './serializationUtils'
+export { SerializationContextStatus, NodeWithSerializedNode } from './serialization.types'
+export { serializeDocument } from './serializeDocument'
+export { serializeNodeWithId } from './serializeNode'
+export { serializeAttribute } from './serializeAttribute'
