+++ packages/rum/src/domain/record/serialization/serializeNode.spec.ts
@@ -1,7 +1,6 @@
 import { isIE, noop } from '@datadog/browser-core'
 
 import type { RumConfiguration } from '@datadog/browser-rum-core'
-import { STABLE_ATTRIBUTES, DEFAULT_PROGRAMMATIC_ACTION_NAME_ATTRIBUTE } from '@datadog/browser-rum-core'
 import {
   NodePrivacyLevel,
   PRIVACY_ATTR_NAME,
@@ -10,34 +9,28 @@ import {
   PRIVACY_ATTR_VALUE_MASK,
   PRIVACY_ATTR_VALUE_MASK_USER_INPUT,
 } from '../../../constants'
-import { isAdoptedStyleSheetsSupported } from '../../../../../core/test/specHelper'
 import {
   HTML,
-  AST_ALLOW,
+  generateLeanSerializedDoc,
   AST_HIDDEN,
   AST_MASK,
   AST_MASK_USER_INPUT,
-  generateLeanSerializedDoc,
+  AST_ALLOW,
 } from '../../../../test/htmlAst'
 import type { ElementNode, SerializedNodeWithId, TextNode } from '../../../types'
 import { NodeType } from '../../../types'
 import type { IsolatedDom } from '../../../../../rum-core/test/createIsolatedDom'
 import { createIsolatedDom } from '../../../../../rum-core/test/createIsolatedDom'
-import { MAX_ATTRIBUTE_VALUE_CHAR_LENGTH } from '../privacy'
 import type { ElementsScrollPositions } from '../elementsScrollPositions'
 import { createElementsScrollPositions } from '../elementsScrollPositions'
 import type { ShadowRootCallBack, ShadowRootsController } from '../shadowRootsController'
 import type { WithAdoptedStyleSheets } from '../browser.types'
-import {
-  serializeDocument,
-  serializeNodeWithId,
-  serializeDocumentNode,
-  serializeChildNodes,
-  serializeAttribute,
-  SerializationContextStatus,
-} from './serialize'
-import type { SerializationContext, SerializeOptions } from './serialize'
+import { isAdoptedStyleSheetsSupported } from '../../../../../core/test/specHelper'
+import { serializeDocument } from './serializeDocument'
+import type { SerializationContext, SerializeOptions } from './serialization.types'
+import { SerializationContextStatus } from './serialization.types'
 import { hasSerializedNode } from './serializationUtils'
+import { serializeChildNodes, serializeDocumentNode, serializeNodeWithId } from './serializeNode'
 
 const DEFAULT_CONFIGURATION = {} as RumConfiguration
 
@@ -868,96 +861,6 @@ describe('serializeDocumentNode handles', function testAllowDomTree() {
     it('is serialized correctly', () => {
       const serializedDoc = generateLeanSerializedDoc(HTML, 'allow')
       expect(toJSONObj(serializedDoc)).toEqual(AST_ALLOW)
-    })
-  })
-})
-
-describe('serializeAttribute ', () => {
-  beforeEach(() => {
-    if (isIE()) {
-      pending('IE not supported')
-    }
-  })
-
-  it('truncates "data:" URIs after long string length', () => {
-    const node = document.createElement('p')
-
-    const longString = new Array(MAX_ATTRIBUTE_VALUE_CHAR_LENGTH + 1 - 5).join('a')
-    const maxAttributeValue = `data:${longString}`
-    const exceededAttributeValue = `data:${longString}1`
-    const ignoredAttributeValue = `foos:${longString}`
-
-    node.setAttribute('test-okay', maxAttributeValue)
-    node.setAttribute('test-truncate', exceededAttributeValue)
-    node.setAttribute('test-ignored', ignoredAttributeValue)
-
-    expect(serializeAttribute(node, NodePrivacyLevel.ALLOW, 'test-okay', DEFAULT_CONFIGURATION)).toBe(maxAttributeValue)
-    expect(serializeAttribute(node, NodePrivacyLevel.MASK, 'test-okay', DEFAULT_CONFIGURATION)).toBe(maxAttributeValue)
-
-    expect(serializeAttribute(node, NodePrivacyLevel.MASK, 'test-ignored', DEFAULT_CONFIGURATION)).toBe(
-      ignoredAttributeValue
-    )
-
-    expect(serializeAttribute(node, NodePrivacyLevel.ALLOW, 'test-truncate', DEFAULT_CONFIGURATION)).toBe(
-      'data:truncated'
-    )
-    expect(serializeAttribute(node, NodePrivacyLevel.MASK, 'test-truncate', DEFAULT_CONFIGURATION)).toBe(
-      'data:truncated'
-    )
-  })
-
-  it('does not mask the privacy attribute', () => {
-    const node = document.createElement('div')
-    node.setAttribute(PRIVACY_ATTR_NAME, NodePrivacyLevel.MASK)
-    expect(serializeAttribute(node, NodePrivacyLevel.MASK, PRIVACY_ATTR_NAME, DEFAULT_CONFIGURATION)).toBe('mask')
-  })
-
-  it('masks data attributes', () => {
-    const node = document.createElement('div')
-    node.setAttribute('data-foo', 'bar')
-    expect(serializeAttribute(node, NodePrivacyLevel.MASK, 'data-foo', DEFAULT_CONFIGURATION)).toBe('***')
-  })
-
-  describe('attributes used to generate CSS selectors', () => {
-    it('does not mask the default programmatic action name attributes', () => {
-      const node = document.createElement('div')
-      node.setAttribute(DEFAULT_PROGRAMMATIC_ACTION_NAME_ATTRIBUTE, 'foo')
-      expect(
-        serializeAttribute(
-          node,
-          NodePrivacyLevel.MASK,
-          DEFAULT_PROGRAMMATIC_ACTION_NAME_ATTRIBUTE,
-          DEFAULT_CONFIGURATION
-        )
-      ).toBe('foo')
-    })
-
-    it('does not mask the user-supplied programmatic action name attributes when it is a data attribute', () => {
-      const node = document.createElement('div')
-      node.setAttribute('data-my-custom-action-name', 'foo')
-      expect(
-        serializeAttribute(node, NodePrivacyLevel.MASK, 'data-my-custom-action-name', {
-          ...DEFAULT_CONFIGURATION,
-          actionNameAttribute: 'data-my-custom-action-name',
-        })
-      ).toBe('foo')
-    })
-
-    it('does not mask the user-supplied programmatic action name attributes when it not a data attribute', () => {
-      const node = document.createElement('div')
-      node.setAttribute('my-custom-action-name', 'foo')
-      expect(
-        serializeAttribute(node, NodePrivacyLevel.MASK, 'my-custom-action-name', {
-          ...DEFAULT_CONFIGURATION,
-          actionNameAttribute: 'my-custom-action-name',
-        })
-      ).toBe('foo')
-    })
-
-    it('does not mask other attributes used to generate CSS selectors', () => {
-      const node = document.createElement('div')
-      node.setAttribute(STABLE_ATTRIBUTES[0], 'foo')
-      expect(serializeAttribute(node, NodePrivacyLevel.MASK, STABLE_ATTRIBUTES[0], DEFAULT_CONFIGURATION)).toBe('foo')
     })
   })
 })
