--- packages/rum/src/domain/record/serializers/index.ts
@@ -1,8 +1,0 @@
-export {
-  getElementInputValue,
-  getSerializedNodeId,
-  hasSerializedNode,
-  nodeAndAncestorsHaveSerializedNode,
-  NodeWithSerializedNode,
-} from './serializationUtils'
-export { serializeDocument, SerializationContextStatus, serializeNodeWithId, serializeAttribute } from './serialize'
